Generalize Favorite into Event Events encapsulate any action that a user can take. An Event instance has an initiating user, an action descriptor, a subject (Toot or User on which the action was performed), and a creation date for ordering events.
This generalization also means that `User#toot_stream` will be generalized to `User#event_stream`, returning the list of all events that the User initiated.

Additionally, there will be an `Event.stream_for_user` method, which will return all events that a User is involved in, either as the initiator, or as the subject. Eventually, this will also include events involving a Users followed users.