Initial @mentioning functionality @mentions are now mostly supported. Templates can use the `with_mentions` helper to scan a block of text and replace any @mentions with links to the respective user profiles.
However, the implementation is very poor, and would be much better done by a simple JS script. This, though, requires that profile URLs be fully determinable from a user's handle, which it currently is not (URLs require User ids). Attempting to make vanity URLs that use the user's handle instead was successful, but ended up regressing a number of tests, so the functionality was dropped for now.